+++ src/plugins/lua/api/text/template_string.cpp
@@ -1,0 +1,18 @@
+#include "beez/plugin/lua/api/text/template_string.hpp"
+
+#include <string>
+
+// NOLINTBEGIN(misc-include-cleaner,cppcoreguidelines-pro-bounds-avoid-unchecked-container-access)
+#include <sol/sol.hpp>
+// NOLINTEND(misc-include-cleaner,cppcoreguidelines-pro-bounds-avoid-unchecked-container-access)
+
+namespace beez::plugin::lua
+{
+
+void bindTemplateString(sol::table& textTable)
+{
+    textTable["template"] = [](const std::string& templateString, const sol::table& /*variables*/) -> std::string
+    { return templateString; };
+}
+
+}  // namespace beez::plugin::lua